+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.7.12-alpha.23 (2025-10-30)</small>
+
+- added streaming functionality ([7f310fc](https://github.com/olane-labs/olane/commit/7f310fc))
+- added timeouts for read + drain ([cad9348](https://github.com/olane-labs/olane/commit/cad9348))
+- updated gitignore ([9f7b095](https://github.com/olane-labs/olane/commit/9f7b095))
+
 ## <small>0.7.12-alpha.22 (2025-10-29)</small>
 
 - fix ([e87ca71](https://github.com/olane-labs/olane/commit/e87ca71))