+++ packages/o-gateway-olane/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.7.12-alpha.23 (2025-10-30)</small>
+
+**Note:** Version bump only for package @olane/o-gateway-olane
+
 ## <small>0.7.12-alpha.22 (2025-10-29)</small>
 
 **Note:** Version bump only for package @olane/o-gateway-olane